Atualizar arquivo UsuarioController.php para ser o controller do usuário. Deletar arquivo UserController.php
